+++ HAREWARE/tsl2561.h
@@ -162,7 +162,7 @@ tsl2561_Error_t;
 void TSL2561_Init(I2C_HandleTypeDef *hi2c);
 void TSL2561_SetTiming(I2C_HandleTypeDef *hi2c, tsl2561_IntegrationTime_t integration, tsl2561_Gain_t gain);
 void TSL2561_GetLuminosity (I2C_HandleTypeDef *hi2c, uint16_t *broadband, uint16_t *ir);
-uint32_t TSL2561_CalculateLux(uint16_t ch0, uint16_t ch1);
+uint16_t TSL2561_CalculateLux(uint16_t ch0, uint16_t ch1);
 
 #endif
